+++ src/api/controllers/search/index.js
@@ -1,4 +1,9 @@
-const parseAccountName = require('lib/user').parseAccountName;
+function parseAccountName (userResult) {
+  const token = userResult.gw2_api_tokens &&
+    userResult.gw2_api_tokens.filter((tkn) => tkn.primary)[0];
+
+  return token && token.accountName;
+}
 
 function controller (models) {
   function search (term) {