+++ src/lib/services/user.spec.js
@@ -1,3 +1,4 @@
+import _ from 'lodash';
 import * as testData from 'test/testData';
 
 const readGuild = sinon.stub();
@@ -70,45 +71,69 @@ describe('user service', () => {
     });
   });
 
-  it('should get user id by email', async () => {
-    const id = await service.getUserIdByEmail(models, user.email);
-    expect(id).to.exist;
-  });
-
-  it('should reject if user doesnt exist by email', async () => {
-    try {
-      await service.getUserIdByEmail(models, 'dont_exist');
-    } catch (e) {
-      expect(e).to.equal('User not found');
-    }
-  });
-
-  it('should get user id by alias', async () => {
-    const id = await service.getUserIdByAlias(models, user.alias);
-    expect(id).to.exist;
-  });
-
-  it('should return user primary token', async () => {
-    const token = await service.getUserPrimaryToken(models, userTwo.alias);
-    expect(token).to.equal(apiTokenForUserTwo.token);
-  });
-
-  it('should return error if no primary token found', async () => {
-    try {
-      await service.getUserPrimaryToken(models, user.alias);
-    } catch (e) {
-      expect(e).to.equal('Token not found');
-    }
-  });
-
-  describe('reading', () => {
-    it('should read user by apiToken', async () => {
-      const usr = await service.read(models, { apiToken: apiToken.token });
-
+  describe('read', () => {
+    const assertUser = (usr) => {
       expect(usr).to.eql({
-        accountName: apiToken.accountName,
-        alias: user.alias,
-        apiToken: apiToken.token,
+        id: userTwo.id,
+        alias: userTwo.alias,
+        email: userTwo.email,
+        passwordHash: userTwo.passwordHash,
+        ..._.pick(apiTokenForUserTwo, [
+          'accountName',
+          'world',
+          'access',
+          'commander',
+          'fractalLevel',
+          'dailyAp',
+          'monthlyAp',
+          'wvwRank',
+          'guilds',
+        ]),
+      });
+    };
+
+    context('with api token', () => {
+      it('should return data', async () => {
+        const usr = await service.read(models, { apiToken: apiTokenForUserTwo.token });
+
+        assertUser(usr);
+      });
+
+      context('when user doesnt exist', () => {
+        it('should return null', async () => {
+          const usr = await service.read(models, { apiToken: 'asd' });
+          expect(usr).to.be.null;
+        });
+      });
+    });
+
+    context('with alias', () => {
+      it('should return data', async () => {
+        const usr = await service.read(models, { alias: userTwo.alias });
+
+        assertUser(usr);
+      });
+
+      context('when user doesnt exist', () => {
+        it('should return null', async () => {
+          const usr = await service.read(models, { alias: 'asd' });
+          expect(usr).to.be.null;
+        });
+      });
+    });
+
+    context('with email', () => {
+      it('should return data', async () => {
+        const usr = await service.read(models, { email: userTwo.email });
+
+        assertUser(usr);
+      });
+
+      context('when user doesnt exist', () => {
+        it('should return null', async () => {
+          const usr = await service.read(models, { email: 'asd' });
+          expect(usr).to.be.null;
+        });
       });
     });
   });