+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - `gw2aRank` and `naRank` to `PvpStandings` table
 - Pvp leaderboard fetcher to calculate gw2aRank and save it to the database
 
+### Changed
+- User service, controller major refactors
+
 ## [1.2.2] - 2017/01/08
 ### Fixed
 - Sitemap to not have empty `lastmod` fields